requirements.txt
streamlit==1.32.0
ultralytics==8.1.0
opencv-python==4.9.0.80
pillow==10.1.0
numpy==1.26.0